+++ ReleaseNotes.md
@@ -6,5 +6,6 @@
 
 ## 1.0.1
 
+- Added symbol files to NuGet package if there (doesn't need AddSymbols to be on)
 - Added copy of symbol files to NuGet cache when using the U(pdate) command.